requirements.txt
anyio==4.12.0
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
orjson==3.11.5
packaging==25.0
pillow==12.1.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.1
pyparsing==3.3.1
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
six==1.17.0
sympy==1.14.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
typing_extensions==4.15.0
ultralytics==8.3.248
ultralytics-thop==2.0.18
urllib3==2.6.2
